requirements.txt
pyperclip
pillow
pynput
